flytt fradragsgrunnlag extension funs inn i Fradragsgrunnlag (Grunnlag.kt)
